// src/store/entities/enums/store.enum.ts
export enum StoreEnum {
  // Theme
  THEME = 'theme',
  LIGHT = 'light',
  DARK = 'dark',

  // Language
  LANGUAGE = 'language',
  LANGUAGE_EN = 'en',
  LANGUAGE_ES = 'es',

  // Icons
  SUMMER = 'summer',
  WINTER = 'winter',
  SPRING = 'spring',
  AUTUMN = 'autumn',

  // User
  STATUS_LOADING = 'loading',
  STATUS_AUTHENTICATED = 'authenticated',
  STATUS_UNAUTHENTICATED = 'unauthenticated',
  TOKEN = 'token',
}