// src/common/exceptions/entities/enums/settings-exception.enum.ts
export enum SettingsExceptionMessages {
  FetchSettingsException = "Error fetching settings:",
  UpdateSettingsException = "Error updating settings:",
  UserIDRequiredException = "User ID is required to fetch settings.",
}

export enum SettingsExceptionNames {
  FetchSettingsException = "FetchSettingsException",
  UpdateSettingsException = "UpdateSettingsException",
  UserIDRequiredException = "UserIDRequiredException",
}